server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
Werkzeug==2.0.1
tensorflow==2.8.0
numpy==1.22.3
pymongo==4.1.1
python-dotenv==0.19.2